Socket.io
Added some socket testing and moved some files from client to shared.
