+++ Project-MurderDome/client/public/javascripts/shared/src/classes/Player.js
@@ -1,100 +1,65 @@
-
-import type { action } from '../types/types';
-import { Action } from './Action';
-
+import { Action } from './Action.js';
 export class Player {
-
-    readonly parent: HTMLDivElement;
-    readonly name: string;
-
-    private _container: HTMLDivElement;
-    private _selectElement: HTMLSelectElement;
-    private _changeEventHandler: EventListener;
-    private _selectedAction: Action;
-
-    constructor(parentElement: HTMLDivElement, playerName: string) {
-
+    constructor(parentElement, playerName) {
         this.parent = parentElement;
         this.name = playerName;
-
         this.parent.classList.add("PlayerControl");
-
         this._container = document.createElement("div");
         this._container.classList.add("PlayerContainer");
-
         this._container.appendChild(this._createLabel());
         this._container.appendChild(this._createSelect());
-
         this.parent.appendChild(this._container);
-
-        return
+        return;
     }
-
-
-    private _createLabel(): HTMLLabelElement {
-
-        let label: HTMLLabelElement = document.createElement('label');
+    _createLabel() {
+        let label = document.createElement('label');
         label.classList.add("PlayerLabel");
         label.setAttribute('for', this.name);
         label.innerText = this.name;
-
         return label;
     }
-
-    private _createSelect(): HTMLSelectElement {
-
-        let select: HTMLSelectElement = document.createElement('select');
+    _createSelect() {
+        let select = document.createElement('select');
         select.setAttribute('name', this.name);
-
         Action.playerActions.forEach((action) => {
             select.appendChild(this._createOption(action));
         });
-
         this._selectElement = select;
         this._changeEventHandler = () => { this._setSelectedAction(); };
         this._selectElement.addEventListener("change", this._changeEventHandler);
         this._setSelectedAction();
-
         return select;
     }
-
-    private _createOption(action: string): HTMLOptionElement {
-
-        let option: HTMLOptionElement = document.createElement('option');
+    _createOption(action) {
+        let option = document.createElement('option');
         option.setAttribute('value', action);
         option.innerText = action;
-
         return option;
     }
-
-    private _setSelectedAction(): void {
-
+    _setSelectedAction() {
         if (this._selectElement && this._selectElement.selectedIndex != -1 && this._selectElement.options.length > 0) {
-
-            let selectedOption: HTMLOptionElement = this._selectElement.options.item(this._selectElement.selectedIndex);
-
+            let selectedOption = this._selectElement.options.item(this._selectElement.selectedIndex);
             if (Action.isValidAction(selectedOption.value)) {
-                this._selectedAction = new Action(selectedOption.value as action, this.name);
-            } else {
+                this._selectedAction = new Action(selectedOption.value, this.name);
+            }
+            else {
                 this._selectedAction = undefined;
             }
         }
-        return
+        return;
     }
-
     //private _attachChangeEventHandler(): void {
     //    if (this._selectElement && this._changeEventHandler) {
     //        this._selectElement.addEventListener("change", this._changeEventHandler);
     //    }
     //}
-
     //private _detachChangeEventHandler(): void {
     //    if (this._selectElement && this._changeEventHandler) {
     //        this._selectElement.removeEventListener("change", this._changeEventHandler);
     //    }
     //}
-
-    public getSelectedAction(): Action {
+    getSelectedAction() {
         return this._selectedAction;
     }
 }
+//# sourceMappingURL=Player.js.map
