--- Project-MurderDome/client/public/javascripts/types/types.js
@@ -1,1 +1,0 @@
-//# sourceMappingURL=types.js.map
