--- Project-MurderDome/client/public/stylesheets/main.css
@@ -1,9 +1,0 @@
-body {
-    padding: 50px;
-    font: 14px "Lucida Grande", Helvetica, Arial, sans-serif;
-    background-color: red;
-}
-
-a {
-    color: #00B7FF;
-}
