+++ Project-MurderDome/client/public/views/testPage.html
@@ -5,9 +5,8 @@
     <meta charset="utf-8" />
     <title>HTML Test</title>
     <link rel="stylesheet" href="/stylesheets/testStyle.css" />
-    <!--<script data-main="/javascripts/main" src="/javascripts/require.js"></script>-->
-    <!--<script src="/javascripts/require.js"></script>-->
-    <script type="module" src="/javascripts/scripts/testScript.js"></script>
+    <script src="/socket.io/socket.io.js"></script>
+    <script type="module" src="/javascripts/client/src/scripts/testScript.js"></script>
 </head>
     <body>
         <h1>Players:</h1>
@@ -18,8 +17,6 @@
         <div id="Player4"></div>
         <div id="Player5"></div>
         <div id="Player6"></div>
-        <!--<div id="Player7"></div>
-        <div id="Player8"></div>-->
         
         <button id="EnterBtn" type="button">Enter</button>
         <div id="OutputLabel">Output:</div>
