--- Project-MurderDome/client/views/index.pug
@@ -1,5 +1,0 @@
-extends layout
-
-block content
-  h1= title
-  p Welcome to #{title}
