--- Project-MurderDome/client/views/layout.pug
@@ -1,7 +1,0 @@
-doctype html
-html
-  head
-    title= title
-    link(rel='stylesheet', href='/stylesheets/main.css')
-  body
-    block content
