+++ Project-MurderDome/package.json
@@ -13,16 +13,19 @@
   },
   "main": "server.js",
   "dependencies": {
+    "@types/socket.io": "^2.1.6",
+    "@types/socket.io-client": "^1.4.33",
     "debug": "^2.6.9",
     "express": "^4.17.1",
     "html": "^1.0.0",
     "path": "^0.12.7",
     "pug": "^2.0.0-rc.3",
+    "socket.io": "^2.3.0",
     "text": "^0.1.0"
   },
   "devDependencies": {
     "@types/debug": "0.0.30",
-    "@types/express": "^4.0.37",
+    "@types/express": "^4.17.6",
     "@types/express-serve-static-core": "^4.0.50",
     "@types/mime": "^1.3.1",
     "@types/node": "^8.0.14",
