--- Project-MurderDome/server/routes/index.ts
@@ -1,11 +1,0 @@
-/*
- * GET home page.
- */
-import express = require('express');
-const router = express.Router();
-
-router.get('/', (req: express.Request, res: express.Response) => {
-    res.render('index', { title: 'Express' });
-});
-
-export default router;
