--- Project-MurderDome/server/routes/user.ts
@@ -1,11 +1,0 @@
-/*
- * GET users listing.
- */
-import express = require('express');
-const router = express.Router();
-
-router.get('/', (req: express.Request, res: express.Response) => {
-    res.send("respond with a resource");
-});
-
-export default router;
