+++ Project-MurderDome/server/server.js
@@ -3,21 +3,28 @@ Object.defineProperty(exports, "__esModule", { value: true });
 const debug = require("debug");
 const express = require("express");
 const path = require("path");
-const testRoute_1 = require("./routes/testRoute");
-const index_1 = require("./routes/index");
-const user_1 = require("./routes/user");
-const classes_1 = require("./routes/classes");
+//import classes from './routes/classes';
 const app = express();
-// view engine setup
-app.set('views', path.join(__dirname, '../client/views'));
-app.set('view engine', 'pug');
-app.use(express.static(path.join(__dirname, '../client/public')));
-app.use('/test', testRoute_1.default);
-app.use('/index', index_1.default);
-app.use('/users', user_1.default);
-app.use('/\*/classes/', classes_1.default);
+let http = require('http').createServer(app);
+let io = require('socket.io')(http);
+//app.use((req, res, next) => {
+//    console.log(req.url);
+//    next();
+//});
 app.get('/', (req, res) => {
-    res.sendFile(path.join(app.get('views'), 'testPage.html'));
+    res.sendFile(path.join(__dirname, '../client/public/views', 'testPage.html'));
+});
+app.use(express.static(path.join(__dirname, '../client/public')));
+//app.use('/\*/classes/', classes);
+io.on('connection', (socket) => {
+    console.log('a user connected');
+    socket.on('EnterBtnClicked', (msg) => {
+        console.log('message: ' + msg);
+        io.emit('Output', msg + "</br></br>This Came From Server!");
+    });
+    socket.on('disconnect', () => {
+        console.log('user disconnected');
+    });
 });
 // catch 404 and forward to error handler
 app.use((req, res, next) => {
@@ -25,29 +32,18 @@ app.use((req, res, next) => {
     err['status'] = 404;
     next(err);
 });
-// error handlers
-// development error handler
-// will print stacktrace
 if (app.get('env') === 'development') {
     app.use((err, req, res, next) => {
         res.status(err['status'] || 500);
-        res.render('error', {
-            message: err.message,
-            error: err
-        });
+        res.send("respond with a BLANK resource");
     });
 }
-// production error handler
-// no stacktraces leaked to user
 app.use((err, req, res, next) => {
     res.status(err.status || 500);
-    res.render('error', {
-        message: err.message,
-        error: {}
-    });
+    res.send("respond with a BLANK resource");
 });
 app.set('port', process.env.PORT || 3000);
-const server = app.listen(app.get('port'), function () {
+const server = http.listen(app.get('port'), function () {
     debug('Express server listening on port ' + server.address().port);
     console.log("Server listeneing on port" + server.address().port);
 });
