+++ Project-MurderDome/server/server.ts
@@ -2,27 +2,45 @@ import debug = require('debug');
 import express = require('express');
 import path = require('path');
 
-import testRoute from './routes/testRoute';
-import index from './routes/index';
-import users from './routes/user';
-import classes from './routes/classes';
+//import classes from './routes/classes';
 
 const app = express();
+let http = require('http').createServer(app);
+let io = require('socket.io')(http);
 
-// view engine setup
-app.set('views', path.join(__dirname, '../client/views'));
-app.set('view engine', 'pug');
+//app.use((req, res, next) => {
+//    console.log(req.url);
+//    next();
+//});
+
+app.get('/', (req: express.Request, res: express.Response) => {
+    res.sendFile(path.join(__dirname, '../client/public/views', 'testPage.html'));
+});
 
 app.use(express.static(path.join(__dirname, '../client/public')));
 
-app.use('/test',testRoute);
-app.use('/index', index);
-app.use('/users', users);
-app.use('/\*/classes/', classes);
+//app.use('/\*/classes/', classes);
 
-app.get('/', (req: express.Request, res: express.Response) => {
-    res.sendFile(path.join(app.get('views'), 'testPage.html'));
+
+
+
+
+
+io.on('connection', (socket) => {
+    console.log('a user connected');
+
+
+    socket.on('EnterBtnClicked', (msg) => {
+        console.log('message: ' + msg);
+        io.emit('Output', msg + "</br></br>This Came From Server!");
+    });
+
+
+    socket.on('disconnect', () => {
+        console.log('user disconnected');
+    });
 });
+
 
 
 // catch 404 and forward to error handler
@@ -32,33 +50,21 @@ app.use((req, res, next) => {
     next(err);
 });
 
-// error handlers
-
-// development error handler
-// will print stacktrace
 if (app.get('env') === 'development') {
-    app.use((err, req, res, next) => { // eslint-disable-line @typescript-eslint/no-unused-vars
+    app.use((err, req, res, next) => {
         res.status(err['status'] || 500);
-        res.render('error', {
-            message: err.message,
-            error: err
-        });
+        res.send("respond with a BLANK resource");
     });
 }
 
-// production error handler
-// no stacktraces leaked to user
-app.use((err, req, res, next) => { // eslint-disable-line @typescript-eslint/no-unused-vars
+app.use((err, req, res, next) => {
     res.status(err.status || 500);
-    res.render('error', {
-        message: err.message,
-        error: {}
-    });
+    res.send("respond with a BLANK resource");
 });
 
 app.set('port', process.env.PORT || 3000);
 
-const server = app.listen(app.get('port'), function () {
+const server = http.listen(app.get('port'), function () {
     debug('Express server listening on port ' + server.address().port);
     console.log("Server listeneing on port" + server.address().port);
 });
