+++ Project-MurderDome/shared/src/classes/Player.ts
@@ -1,65 +1,100 @@
-import { Action } from './Action';
+
+import type { action } from '../types/types';
+import { Action } from './Action.js';
+
 export class Player {
-    constructor(parentElement, playerName) {
+
+    readonly parent: HTMLDivElement;
+    readonly name: string;
+
+    private _container: HTMLDivElement;
+    private _selectElement: HTMLSelectElement;
+    private _changeEventHandler: EventListener;
+    private _selectedAction: Action;
+
+    constructor(parentElement: HTMLDivElement, playerName: string) {
+
         this.parent = parentElement;
         this.name = playerName;
+
         this.parent.classList.add("PlayerControl");
+
         this._container = document.createElement("div");
         this._container.classList.add("PlayerContainer");
+
         this._container.appendChild(this._createLabel());
         this._container.appendChild(this._createSelect());
+
         this.parent.appendChild(this._container);
-        return;
+
+        return
     }
-    _createLabel() {
-        let label = document.createElement('label');
+
+
+    private _createLabel(): HTMLLabelElement {
+
+        let label: HTMLLabelElement = document.createElement('label');
         label.classList.add("PlayerLabel");
         label.setAttribute('for', this.name);
         label.innerText = this.name;
+
         return label;
     }
-    _createSelect() {
-        let select = document.createElement('select');
+
+    private _createSelect(): HTMLSelectElement {
+
+        let select: HTMLSelectElement = document.createElement('select');
         select.setAttribute('name', this.name);
+
         Action.playerActions.forEach((action) => {
             select.appendChild(this._createOption(action));
         });
+
         this._selectElement = select;
         this._changeEventHandler = () => { this._setSelectedAction(); };
         this._selectElement.addEventListener("change", this._changeEventHandler);
         this._setSelectedAction();
+
         return select;
     }
-    _createOption(action) {
-        let option = document.createElement('option');
+
+    private _createOption(action: string): HTMLOptionElement {
+
+        let option: HTMLOptionElement = document.createElement('option');
         option.setAttribute('value', action);
         option.innerText = action;
+
         return option;
     }
-    _setSelectedAction() {
+
+    private _setSelectedAction(): void {
+
         if (this._selectElement && this._selectElement.selectedIndex != -1 && this._selectElement.options.length > 0) {
-            let selectedOption = this._selectElement.options.item(this._selectElement.selectedIndex);
+
+            let selectedOption: HTMLOptionElement = this._selectElement.options.item(this._selectElement.selectedIndex);
+
             if (Action.isValidAction(selectedOption.value)) {
-                this._selectedAction = new Action(selectedOption.value, this.name);
-            }
-            else {
+                this._selectedAction = new Action(selectedOption.value as action, this.name);
+            } else {
                 this._selectedAction = undefined;
             }
         }
-        return;
+        return
     }
+
     //private _attachChangeEventHandler(): void {
     //    if (this._selectElement && this._changeEventHandler) {
     //        this._selectElement.addEventListener("change", this._changeEventHandler);
     //    }
     //}
+
     //private _detachChangeEventHandler(): void {
     //    if (this._selectElement && this._changeEventHandler) {
     //        this._selectElement.removeEventListener("change", this._changeEventHandler);
     //    }
     //}
-    getSelectedAction() {
+
+    public getSelectedAction(): Action {
         return this._selectedAction;
     }
 }
-//# sourceMappingURL=Player.js.map
